requirements.txt
boto3
fastapi
pydantic
faiss-cpu
numpy
sentence-transformers
requests
PyMuPDF
pyyaml
tqdm
spacy
streamlit
uvicorn
urllib3
